requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Werkzeug>=3.0.0
regex>=2023.12.25
